+++ cydrogen/_kx_n.pyx
@@ -5,8 +5,10 @@ from libc.stdint cimport uint8_t
 from libc.string cimport memcpy
 
 from ._basekey cimport BaseKey
-from ._decls cimport kx_keygen, hydro_kx_keypair, hydro_kx_PUBLICKEYBYTES, hydro_kx_SECRETKEYBYTES, hydro_kx_N_PACKET1BYTES
-from ._decls cimport kx_n_1, kx_n_2
+from ._decls cimport kx_keygen, hydro_kx_keypair
+from ._decls cimport hydro_kx_PUBLICKEYBYTES, hydro_kx_SECRETKEYBYTES, hydro_kx_N_PACKET1BYTES
+from ._decls cimport hydro_kx_KK_PACKET1BYTES, hydro_kx_KK_PACKET2BYTES
+from ._decls cimport kx_n_1, kx_n_2, kx_kk_1, kx_kk_2, kx_kk_3
 from ._exceptions cimport KeyExchangeException
 from ._secretbox cimport SecretBoxKey
 
@@ -27,6 +29,25 @@ cdef class Psk(BaseKey):
 cdef class SessionPair:
     def __repr__(self):
         return f'SessionPair(rx={repr(self.rx)}, tx={repr(self.tx)})'
+
+
+cdef class KxKkClientState:
+    def __init__(self, KxPair client_kp):
+        self.packet1 = b""
+        self.session_pair = None
+        self.client_kp = client_kp
+
+    cpdef client_finish_kx_kk(self, bytes packet2):
+        if packet2 is None:
+            raise ValueError("packet2 cannot be None")
+        if len(packet2) != hydro_kx_KK_PACKET2BYTES:
+            raise ValueError(f"Packet2 must be {hydro_kx_KK_PACKET2BYTES} bytes long")
+        try:
+            rx, tx = kx_kk_3(&self.state, packet2, self.client_kp)
+        except RuntimeError as ex:
+            raise KeyExchangeException("failed to finish key exchange") from ex
+        self.session_pair = SessionPair(rx=SecretBoxKey(rx), tx=SecretBoxKey(tx))
+        return self.session_pair
 
 
 cdef class KxPublicKey:
@@ -151,32 +172,57 @@ cdef class KxPair:
     cpdef secret_key(self):
         return KxSecretKey(self.kp)
 
+    cpdef server_finish_kx_n(self, bytes packet1, Psk psk=None):
+        return server_finish_kx_n(self, packet1, psk)
+
+    cpdef client_init_kx_kk(self, KxPublicKey server_public_key):
+        if server_public_key is None:
+            raise ValueError("Server public key cannot be None")
+        cdef KxKkClientState state = KxKkClientState(self)
+        try:
+            state.packet1 = kx_kk_1(&state.state, server_public_key, self)
+        except RuntimeError as ex:
+            raise KeyExchangeException("failed to generate first packet for key exchange") from ex
+        return state
+
+    cpdef server_process_kx_kk(self, KxPublicKey client_public_key, bytes packet1):
+        if client_public_key is None:
+            raise ValueError("Client public key cannot be None")
+        if packet1 is None:
+            raise ValueError("packet1 cannot be None")
+        if len(packet1) != hydro_kx_KK_PACKET1BYTES:
+            raise ValueError(f"Packet1 must be {hydro_kx_KK_PACKET1BYTES} bytes long")
+        try:
+            rx, tx, packet2 = kx_kk_2(packet1, client_public_key, self)
+        except RuntimeError as ex:
+            raise KeyExchangeException("failed to generate session from packet") from ex
+        return SessionPair(rx=SecretBoxKey(rx), tx=SecretBoxKey(tx)), packet2
+
     @classmethod
     def gen(cls):
         return cls(kx_keygen())
 
 
-cpdef kx_n_gen_session_and_packet(KxPublicKey peer, Psk psk=None):
+cpdef client_init_kx_n(KxPublicKey server_public_key, Psk psk=None):
     # generate the first packet for the key exchange
-    if peer is None:
-        raise ValueError("peer cannot be None")
+    if server_public_key is None:
+        raise ValueError("Server public key cannot be None")
     try:
-        rx, tx, packet1 = kx_n_1(peer, psk)
+        rx, tx, packet1 = kx_n_1(server_public_key, psk)
     except RuntimeError as ex:
         raise KeyExchangeException("failed to generate first packet") from ex
-    cdef SessionPair pair = SessionPair(rx=SecretBoxKey(rx), tx=SecretBoxKey(tx))
-    return pair, packet1
+    return SessionPair(rx=SecretBoxKey(rx), tx=SecretBoxKey(tx)), packet1
 
 
-cpdef kx_n_gen_session_from_packet(KxPair static_kp, bytes packet1, Psk psk=None):
-    if static_kp is None:
+cpdef server_finish_kx_n(KxPair server_kp, bytes packet1, Psk psk=None):
+    if server_kp is None:
         raise ValueError("static key pair cannot be None")
     if packet1 is None:
         raise ValueError("packet1 cannot be None")
     if len(packet1) != hydro_kx_N_PACKET1BYTES:
         raise ValueError(f"Packet1 must be {hydro_kx_N_PACKET1BYTES} bytes long")
     try:
-        rx, tx = kx_n_2(packet1, psk, static_kp)
+        rx, tx = kx_n_2(packet1, psk, server_kp)
     except RuntimeError as ex:
         raise KeyExchangeException("failed to generate session from packet") from ex
     return SessionPair(rx=SecretBoxKey(rx), tx=SecretBoxKey(tx))