+++ docs/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `cydrogen` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- key exchange using the KK variant.
+
 ## v0.0.10 - 2025-06-12
 
 ### Fixed